# frontend/templates/drf-yasg/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger Diploma Project",
    "description": "",
    "termsOfService": "https://www.google.com/policies/terms/",
    "contact": {
      "email": "contact@snippets.local"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "host": "127.0.0.1",
  "schemes": [
    "http"
  ],
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Basic": []
    }
  ],
  "paths": {
...
    "/profile/": {
      "get": {
        "operationId": "profile_list",
        "description": "Получить профиль пользователя",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The initial index from which to return the results.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "tags": [
          "profile"
        ]
      },
      "post": {
        "operationId": "profile_create",
        "description": "Обновить профиль пользователя",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "tags": [
          "profile"
        ]
      },
      "parameters": []
    },
    "/profile/avatar/": {
      "post": {
        "operationId": "profile_avatar_create",
        "description": "URL загруженного аватара.",
        "parameters": [
          {
            "name": "avatar",
            "in": "formData",
            "required": false,
            "type": "file",
            "x-nullable": true
          },
          {
            "name": "last_name",
            "in": "formData",
            "required": false,
            "type": "string",
            "maxLength": 150
          },
          {
            "name": "first_name",
            "in": "formData",
            "required": false,
            "type": "string",
            "maxLength": 150
          },
          {
            "name": "surname",
            "in": "formData",
            "required": false,
            "type": "string",
            "maxLength": 50
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserAvatar"
            }
          }
        },
        "consumes": [
          "multipart/form-data"
        ],
        "tags": [
          "profile"
        ]
      },
      "parameters": []
    },
...
  },
  "definitions": {
...
    "User": {
      "required": [
        "fullName",
        "email"
      ],
      "type": "object",
      "properties": {
        "fullName": {
          "title": "ФИО",
          "type": "string",
          "maxLength": 100,
          "minLength": 1
        },
        "phone": {
          "title": "Номер телефона",
          "type": "string",
          "maxLength": 20,
          "x-nullable": true
        },
        "email": {
          "title": "Адрес электронной почты",
          "type": "string",
          "format": "email",
          "maxLength": 254,
          "minLength": 1
        },
        "avatar": {
          "title": "Аватар",
          "type": "string",
          "readOnly": true,
          "x-nullable": true,
          "format": "uri"
        }
      }
    },
    "UserAvatar": {
      "type": "object",
      "properties": {
        "avatar": {
          "title": "Аватар",
          "type": "string",
          "readOnly": true,
          "x-nullable": true,
          "format": "uri"
        },
        "last_name": {
          "title": "Фамилия",
          "type": "string",
          "maxLength": 150
        },
        "first_name": {
          "title": "Имя",
          "type": "string",
          "maxLength": 150
        },
        "surname": {
          "title": "Отчество",
          "type": "string",
          "maxLength": 50
        }
      }
    },
...
  }
}